requirements.txt
asgiref==3.2.10
cffi==1.14.2
cryptography==3.1
Django==3.0.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
djongo==1.3.3
Pillow==7.2.0
pycparser==2.20
pymongo==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
